+++ packages/sub-bar/test/all.test.ts
@@ -2,3 +2,4 @@ import "./formatting.test.js";
 import "./settings.test.js";
 import "./dividers.test.js";
 import "./providers.test.js";
+import "./status.test.js";