+++ packages/sub-core/test/all.test.ts
@@ -3,3 +3,4 @@ import "./providers.test.js";
 import "./controller.test.js";
 import "./cache.test.js";
 import "./lock.test.js";
+import "./status.test.js";